+++ src/lib/skilling/skills/crafting/craftables/silver.ts
@@ -1,6 +1,6 @@
-import { Craftable } from '../../types';
-import itemID from '../../../util/itemID';
-import { transformStringBankToNum } from '../../../util/transformStringBankToNum';
+import { Craftable } from '../../../types';
+import itemID from '../../../../util/itemID';
+import { transformStringBankToNum } from '../../../../util/transformStringBankToNum';
 
 const Silver: Craftable[] = [
 	{
