+++ src/tasks/minions/craftingActivity.ts
@@ -5,7 +5,7 @@ import { Time } from '../../lib/constants';
 import { SkillsEnum } from '../../lib/skilling/types';
 import { CraftingActivityTaskOptions } from '../../lib/types/minions';
 import getUsersPerkTier from '../../lib/util/getUsersPerkTier';
-import Crafting from '../../lib/skilling/skills/crafting';
+import Crafting from '../../lib/skilling/skills/crafting/crafting';
 import { channelIsSendable } from '../../lib/util/channelIsSendable';
 
 export default class extends Task {
